Add: 재생 버튼 클릭 시 시간 줄어들도록 설정

+++ app/src/main/java/com/example/plan_me/entity/Time.kt
@@ -4,10 +4,10 @@ import androidx.room.Entity
 import androidx.room.PrimaryKey
 
 @Entity(tableName = "TimeTable")
-data class Time (
-    var focusTime : Int = 50,
-    var breakTime : Int = 10,
-    var repeatCount : Int = 1
+data class Time(
+    var focusTime: Long = 50,
+    var breakTime: Long = 10,
+    var repeatCount: Int = 1
 ) {
     @PrimaryKey var set: Int? = null
 }
